+++ tests/core/behaviors/JsonBehaviorTest.php
@@ -12,7 +12,7 @@ class JsonBehaviorTest extends LuyaWebTestCase
     public function testEncodingDecodingOfAttributes()
     {
         $model = new DynamicModel(['json' => null]);
-        $model->attachBehavior('jsonBehavior', ['class' => JsonBehavior::class, 'attributes' => ['json'], 'decodeAfterFind' => true]);
+        $model->attachBehavior('jsonBehavior', ['class' => JsonBehavior::class, 'attributes' => ['json'], 'decodeAfterFind' => true, 'encodeBeforeValidate' => true]);
 
         $model->addRule('json', 'string');
 
@@ -68,5 +68,22 @@ class JsonBehaviorTest extends LuyaWebTestCase
 
         $this->assertSame('[1,2]', $model->jsonEncode('[1,2]'));
 
+    }
+
+    public function testExpectedDefaultBehavior()
+    {
+        $model = new DynamicModel(['json' => null]);
+        $model->attachBehavior('jsonBehavior', ['class' => JsonBehavior::class, 'attributes' => ['json']]);
+        $model->addRule('json', 'each', ['rule' => ['safe']]);
+
+        $model->json = [1,2,3];
+
+        $this->assertTrue($model->validate());
+
+        $this->assertSame('[1,2,3]', $model->json);
+
+        $model->trigger(ActiveRecord::EVENT_AFTER_FIND);
+
+        $this->assertSame([1,2,3], $model->json);
     }
 }